requirements.txt
flask>=2.3.0
python-dotenv>=1.0.0
requests>=2.31.0
Pillow>=10.0.0
openai-whisper>=20231117
groq>=0.4.2
playwright>=1.40.0
google-api-python-client>=2.108.0
google-auth-oauthlib>=1.2.0
google-auth-httplib2>=0.2.0
schedule>=1.2.1
